+++ apps/cowswap-frontend/src/modules/tradeFormValidation/services/validateTradeForm.ts
@@ -32,6 +32,8 @@ export function validateTradeForm(context: TradeFormValidationContext): TradeFor
     isRestrictedForCountry,
     isBalancesLoading,
     isBundlingSupported,
+    injectedWidgetParams,
+    tradePriceImpact,
   } = context
 
   const {
@@ -178,6 +180,32 @@ export function validateTradeForm(context: TradeFormValidationContext): TradeFor
 
   if (isWrapUnwrap) {
     validations.push(TradeFormValidation.WrapUnwrapFlow)
+  }
+
+  const { whenPriceImpactIsHigherThan, whenPriceImpactIsUnknown } = injectedWidgetParams?.disableTrade || {}
+
+  const checkHighPriceImpact = typeof whenPriceImpactIsHigherThan === 'number'
+  const checkUnknownPriceImpact = whenPriceImpactIsUnknown || checkHighPriceImpact
+
+  if (!isWrapUnwrap && checkUnknownPriceImpact) {
+    const isPriceImpactUnknown = !tradePriceImpact.loading && !tradePriceImpact.priceImpact
+
+    if (isPriceImpactUnknown) {
+      validations.push(TradeFormValidation.DisableTradeWithUnknownPriceImpact)
+    }
+
+    if (tradePriceImpact.loading) {
+      validations.push(TradeFormValidation.ImpactLoading)
+    }
+  }
+
+  if (!isWrapUnwrap && checkHighPriceImpact && tradePriceImpact.priceImpact) {
+    const priceImpactAsNum = +tradePriceImpact.priceImpact.toSignificant()
+    const isPriceImpactAboveThreshold = priceImpactAsNum > whenPriceImpactIsHigherThan
+
+    if (isPriceImpactAboveThreshold) {
+      validations.push(TradeFormValidation.DisableTradeWithHighPriceImpact)
+    }
   }
 
   if (![ApproveRequiredReason.Unsupported, ApproveRequiredReason.NotRequired].includes(isApproveRequired)) {