+++ apps/widget-configurator/src/app/configurator/utils/validateURL.ts
@@ -1,14 +1,5 @@
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export const validateURL = (url: string) => {
-  const pattern = new RegExp(
-    '^(https?:\\/\\/)?' + // protocol
-      '((([a-z\\d]([a-z\\d-]*[a-z\\d])*)\\.)+[a-z]{2,}|' + // domain name
-      '((\\d{1,3}\\.){3}\\d{1,3}))' + // OR ip (v4) address
-      '(\\:\\d+)?(\\/[-a-z\\d%_.~+]*)*' + // port and path
-      '(\\?[;&a-z\\d%_.~+=-]*)?' + // query string
-      '(\\#[-a-z\\d_]*)?$',
-    'i',
-  ) // fragment locator
-  return !!pattern.test(url)
+import { parseENSAddress, uriToHttp } from '@cowprotocol/common-utils'
+
+export const validateURL = (url: string): boolean => {
+  return uriToHttp(url).length > 0 || Boolean(parseENSAddress(url))
 }